requirements.txt
mysql-connector-python==8.0.33
pytest==7.4.0
pytest-mock==3.11.1
requests==2.31.0
sseclient-py==1.8.0